backend/requirements.txt
fastapi==0.128.4

uvicorn==0.40.0
uvicorn[standard]==0.40.0

pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0

python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

duckdb>=0.10.0
python-dotenv>=1.0.1
